#609 Rename quetoo-update to quetoo-launcher.
