--- linux/quetoo/Quetoo Update.desktop
@@ -1,9 +1,0 @@
-[Desktop Entry]
-Name=Quetoo BETA Update
-Comment=Update utility for Quetoo BETA.
-Exec="%QUETOO_HOME%/bin/quetoo-update"
-Icon="%QUETOO_HOME%/share/default/icons/update.png"
-Terminal=true
-Type=Application
-Categories=Games;
-StartupNotify=true
