--- linux/quetoo/README.txt
@@ -1,18 +1,0 @@
-Quetoo BETA - http://quetoo.org
-
-Thank you for downloading Quetoo BETA for Linux. In order to run the game,
-you must first run ./bin/update.sh. This will retrieve game data and updated 
-binaries from the official Quetoo release server.
-
-The first time you run ./bin/update.sh, it may take several minutes. However,
-on subsequent runs, it should be significantly faster, as only new or changed
-content will be downloaded.
-
-You should run ./bin/update.sh regularly, as the game is constantly being added
-to and improved.
-
-Once ./bin/update.sh completes, you may run ./bin/quetoo. You can also link
-the .desktop items for Quetoo and Quetoo Update to the appropriate location(s)
-for your distribution and desktop environment.
-
-Questions or comments? Please join us in #quetoo on irc.freenode.net.
