+++ src/tools/quetoo-launcher/quetoo-launcher-icon.rc
@@ -1,0 +1,1 @@
+update ICON "quetoo-launcher.ico"
