+++ src/tools/quetoo-launcher/quetoo-launcher.ico
