--- src/tools/quetoo-update/quetoo-update-icon.rc
@@ -1,1 +1,0 @@
-update ICON "quetoo-update.ico"
